Implement WebTileImager
This is used for a novel demonstration of downlinking satellite imagery
from a payload through the Antaris Cloud Platform. This just uses a
public web tile dataset for convenience.

For now this just generates a point to use for image collection. It will
be extended to use the payload interface to grab position and attitude
information when available.